requirements.txt
feedparser==6.0.11
yfinance==0.2.51
requests==2.32.3
beautifulsoup4==4.12.3
openai>=1.0.0
python-telegram-bot==21.9
aiosmtplib==3.0.2
python-dotenv==1.0.1
schedule==1.2.2
jinja2==3.1.4
loguru==0.7.3
pytz==2024.2
markdown2>=2.5.0
